+++ src/api/schemas/info.js
@@ -1,0 +1,10 @@
+const InfoSchema = {
+  type: 'object',
+  properties: {
+    content: {type: 'string'}
+  },
+  required: ['content'],
+  additionalProperties: false
+}
+
+module.exports = InfoSchema
