+++ src/api/schemas/movement.js
@@ -1,0 +1,11 @@
+const MovementSchema = {
+  type: 'object',
+  properties: {
+    order: {type: 'integer'},
+    name: {type: 'string'}
+  },
+  required: ['order', 'name'],
+  additionalProperties: false
+}
+
+module.exports = MovementSchema
